Created types for Status and Work component
