+++ src/types/LinkProps.ts
@@ -1,0 +1,4 @@
+export interface LinkProps {
+  href: string;
+  children?: React.ReactNode;
+}
